change worker chunking to respect chunkspec

+++ src/server/worker/driver/testing.go
@@ -148,8 +148,8 @@ func (md *MockDriver) PachClient() *client.APIClient {
 }
 
 // GetExpectedNumWorkers returns the configured number of workers
-func (md *MockDriver) GetExpectedNumWorkers() (int, error) {
-	return md.options.NumWorkers, nil
+func (md *MockDriver) GetExpectedNumWorkers() (uint64, error) {
+	return uint64(md.options.NumWorkers), nil
 }
 
 // NumShards returns the number of hashtree shards configured for the pipeline
